+++ barretenberg/cpp/src/barretenberg/plonk_honk_shared/execution_trace/execution_trace_usage_tracker.hpp
@@ -33,16 +33,17 @@ struct ExecutionTraceUsageTracker {
     size_t max_tables_size = 0;
 
     // For printing only. Must match the order of the members in the arithmetization
+
     static constexpr std::array<std::string_view, 13> block_labels{ "ecc_op",
-                                                                    "pub_inputs",
                                                                     "busread",
+                                                                    "lookup",
+                                                                    "pub_inputs",
                                                                     "arithmetic",
                                                                     "delta_range",
                                                                     "elliptic",
                                                                     "aux",
                                                                     "poseidon2_external",
                                                                     "poseidon2_internal",
-                                                                    "lookup",
                                                                     "overflow",
                                                                     "databus_table_data",
                                                                     "lookup_table_data" };
@@ -88,21 +89,18 @@ struct ExecutionTraceUsageTracker {
         }
 
         // The active ranges must also include the rows where the actual databus and lookup table data are stored.
-        // (Note: lookup tables are constructed at the end of the trace; databus data is constructed at the start) so we
-        // need to determine the dyadic size for this. We call the size function on the current circuit which will have
-        // the same fixed block sizes but might also have an overflow block potentially influencing the dyadic circuit
-        // size.
-        // TODO(https://github.com/AztecProtocol/barretenberg/issues/1160)
-        const size_t dyadic_circuit_size = circuit.blocks.get_structured_dyadic_size();
+        // (Note: lookup tables are constructed from the beginning of the lookup block ; databus data is constructed at
+        // the start of the trace).
 
         // TODO(https://github.com/AztecProtocol/barretenberg/issues/1152): should be able to use simply Range{ 0,
-        // max_databus_size } but this breaks for certain choices of num_threads.
+        // max_databus_size } but this breaks for certain choices of num_threads. It should also be possible to have the
+        // lookup table data be Range{lookup_start, max_tables_size} but that also breaks.
         size_t databus_end =
             std::max(max_databus_size, static_cast<size_t>(fixed_sizes.busread.trace_offset + max_sizes.busread));
         active_ranges.push_back(Range{ 0, databus_end });
-        size_t lookups_start =
-            std::min(dyadic_circuit_size - max_tables_size, static_cast<size_t>(fixed_sizes.lookup.trace_offset));
-        active_ranges.push_back(Range{ lookups_start, dyadic_circuit_size });
+        size_t lookups_start = fixed_sizes.lookup.trace_offset;
+        size_t lookups_end = lookups_start + std::max(max_tables_size, static_cast<size_t>(max_sizes.lookup));
+        active_ranges.emplace_back(Range{ lookups_start, lookups_end });
     }
 
     // Check whether an index is contained within the active ranges (or previous active ranges; needed for perturbator)