+++ yarn-project/end-to-end/src/e2e_epochs.test.ts
@@ -24,24 +24,27 @@ describe('e2e_epochs', () => {
   let l1BlockNumber: number;
   let handle: NodeJS.Timeout;
 
-  const EPOCH_DURATION = 4;
-  const L1_BLOCK_TIME = 5;
-  const L2_SLOT_DURATION_IN_L1_BLOCKS = 2;
+  const EPOCH_DURATION_IN_L2_SLOTS = 4;
+  const L2_SLOT_DURATION_IN_L1_SLOTS = 2;
+  const L1_BLOCK_TIME_IN_S = process.env.L1_BLOCK_TIME ? parseInt(process.env.L1_BLOCK_TIME) : 8;
 
-  beforeAll(async () => {
+  beforeEach(async () => {
     // Set up system without any account nor protocol contracts
     // and with faster block times and shorter epochs.
     context = await setup(0, {
       assumeProvenThrough: undefined,
       skipProtocolContracts: true,
       salt: 1,
-      aztecEpochDuration: EPOCH_DURATION,
-      aztecSlotDuration: L1_BLOCK_TIME * L2_SLOT_DURATION_IN_L1_BLOCKS,
-      ethereumSlotDuration: L1_BLOCK_TIME,
-      aztecEpochProofClaimWindowInL2Slots: EPOCH_DURATION / 2,
+      aztecEpochDuration: EPOCH_DURATION_IN_L2_SLOTS,
+      aztecSlotDuration: L1_BLOCK_TIME_IN_S * L2_SLOT_DURATION_IN_L1_SLOTS,
+      ethereumSlotDuration: L1_BLOCK_TIME_IN_S,
+      aztecEpochProofClaimWindowInL2Slots: EPOCH_DURATION_IN_L2_SLOTS / 2,
       minTxsPerBlock: 0,
       realProofs: false,
       startProverNode: true,
+      // This must be enough so that the tx from the prover is delayed properly,
+      // but not so much to hang the sequencer and timeout the teardown
+      txPropagationMaxQueryAttempts: 12,
     });
 
     logger = context.logger;
@@ -87,16 +90,17 @@ describe('e2e_epochs', () => {
 
     // Constants used for time calculation
     constants = {
-      epochDuration: EPOCH_DURATION,
-      slotDuration: L1_BLOCK_TIME * L2_SLOT_DURATION_IN_L1_BLOCKS,
+      epochDuration: EPOCH_DURATION_IN_L2_SLOTS,
+      slotDuration: L1_BLOCK_TIME_IN_S * L2_SLOT_DURATION_IN_L1_SLOTS,
       l1GenesisBlock: await rollup.getL1StartBlock(),
       l1GenesisTime: await rollup.getL1GenesisTime(),
+      ethereumSlotDuration: L1_BLOCK_TIME_IN_S,
     };
 
     logger.info(`L2 genesis at L1 block ${constants.l1GenesisBlock} (timestamp ${constants.l1GenesisTime})`);
   });
 
-  afterAll(async () => {
+  afterEach(async () => {
     clearInterval(handle);
     await context.teardown();
   });
@@ -105,13 +109,18 @@ describe('e2e_epochs', () => {
   const waitUntilEpochStarts = async (epoch: number) => {
     const [start] = getTimestampRangeForEpoch(BigInt(epoch), constants);
     logger.info(`Waiting until L1 timestamp ${start} is reached as the start of epoch ${epoch}`);
-    await waitUntilL1Timestamp(l1Client, start - BigInt(L1_BLOCK_TIME));
+    await waitUntilL1Timestamp(l1Client, start - BigInt(L1_BLOCK_TIME_IN_S));
     return start;
   };
 
   /** Waits until the given L2 block number is mined. */
   const waitUntilL2BlockNumber = async (target: number) => {
     await retryUntil(() => Promise.resolve(target === l2BlockNumber), `Wait until L2 block ${target}`, 60, 0.1);
+  };
+
+  /** Waits until the given L2 block number is marked as proven. */
+  const waitUntilProvenL2BlockNumber = async (target: number) => {
+    await retryUntil(() => Promise.resolve(target === l2ProvenBlockNumber), `Wait proven L2 block ${target}`, 60, 0.1);
   };
 
   it('does not allow submitting proof after epoch end', async () => {
@@ -124,12 +133,14 @@ describe('e2e_epochs', () => {
     proverDelayer.pauseNextTxUntilTimestamp(epoch2Start);
     logger.info(`Delayed prover tx until epoch 2 starts at ${epoch2Start}`);
 
-    // Wait until the last block of epoch 1 is published and then hold off the sequencer
-    await waitUntilL2BlockNumber(blockNumberAtEndOfEpoch0 + EPOCH_DURATION);
-    sequencerDelayer.pauseNextTxUntilTimestamp(epoch2Start + BigInt(L1_BLOCK_TIME));
+    // Wait until the last block of epoch 1 is published and then hold off the sequencer.
+    // Note that the tx below will block the sequencer until it times out
+    // the txPropagationMaxQueryAttempts until #10824 is fixed.
+    await waitUntilL2BlockNumber(blockNumberAtEndOfEpoch0 + EPOCH_DURATION_IN_L2_SLOTS);
+    sequencerDelayer.pauseNextTxUntilTimestamp(epoch2Start + BigInt(L1_BLOCK_TIME_IN_S));
 
     // Next sequencer to publish a block should trigger a rollback to block 1
-    await waitUntilL1Timestamp(l1Client, epoch2Start + BigInt(L1_BLOCK_TIME));
+    await waitUntilL1Timestamp(l1Client, epoch2Start + BigInt(L1_BLOCK_TIME_IN_S));
     expect(await rollup.getBlockNumber()).toEqual(1n);
     expect(await rollup.getSlotNumber()).toEqual(8n);
 
@@ -142,5 +153,17 @@ describe('e2e_epochs', () => {
     const lastL2BlockTxReceipt = await l1Client.getTransactionReceipt({ hash: lastL2BlockTxHash! });
     expect(lastL2BlockTxReceipt.status).toEqual('success');
     expect(lastL2BlockTxReceipt.blockNumber).toBeGreaterThan(lastProverTxReceipt!.blockNumber);
+    logger.info(`Test succeeded`);
+  });
+
+  it('submits proof claim alone if there is no txs to build a block', async () => {
+    context.sequencer?.updateSequencerConfig({ minTxsPerBlock: 1 });
+    await waitUntilEpochStarts(1);
+    const blockNumberAtEndOfEpoch0 = Number(await rollup.getBlockNumber());
+    logger.info(`Starting epoch 1 after L2 block ${blockNumberAtEndOfEpoch0}`);
+
+    await waitUntilProvenL2BlockNumber(blockNumberAtEndOfEpoch0);
+    expect(l2BlockNumber).toEqual(blockNumberAtEndOfEpoch0);
+    logger.info(`Test succeeded`);
   });
 });